+++ example/lib/demos/node_canvas_clean/layers/background_grid_layer.dart
@@ -2,6 +2,6 @@ import 'package:flutter/material.dart';
 
 import '../painters.dart';
 
-CustomPainter buildBackgroundGridPainter(Matrix4 transform) {
+CustomPainter buildBackgroundGridPainter(Matrix4 transform, _) {
   return InfiniteGridPainter(transform: transform);
 }